instruments & universe

+++ src/rvs.AlgoTrader.Application/DTOs/Orders/OrderDto.cs
@@ -2,7 +2,7 @@ namespace rvs.AlgoTrader.Application.DTOs.Orders;
 
 public record OrderDto(
     Guid Id, string BrokerName, string? BrokerOrderId, string InternalSymbol,
-    string OrderType, string Direction, int Quantity, decimal? Price, decimal? TriggerPrice,
+    string OrderType, string Direction, long Quantity, decimal? Price, decimal? TriggerPrice,
     string Status, Guid? StrategyRunId, DateTimeOffset? PlacedAt, DateTimeOffset? FilledAt,
     decimal? FillPrice, decimal? TrailingSl, decimal? TrailingTp, DateTimeOffset CreatedAt);
